move built-in algorithm class to algorithm module
