+++ python/pyraphtory/pyraphtory/algorithm.py
@@ -1,4 +1,5 @@
 from pyraphtory.graph import TemporalGraph, Row, Table
+from pyraphtory.interop import ScalaProxyBase, find_class, to_python
 
 
 class PyAlgorithm(object):
@@ -24,3 +25,22 @@ class PyAlgorithm(object):
         :return: Table with output from this algorithm
         """
         return graph.global_select(lambda s: Row())
+
+
+class BuiltinAlgorithm(ScalaProxyBase):
+    """Proxy object for looking up built-in algorithms based on path
+
+    (This actually could be used for looking up other classes as well if needed)
+    """
+    @property
+    def _jvm_object(self):
+        return find_class(self._path)
+
+    def __init__(self, path: str):
+        self._path = path
+
+    def __call__(self, *args, **kwargs):
+        return to_python(self._jvm_object).apply(*args, **kwargs)
+
+    def __getattr__(self, item):
+        return BuiltinAlgorithm(".".join((self._path, item)))
